+++ Bytecode/debug/debug.h
@@ -6,12 +6,12 @@
 #include <stdint.h>
 
 void disassembleChunk(Chunk* chunk, const int8_t* name);
-int16_t disassembleInstruction(Chunk* chunk, int16_t offset);
-static int16_t simpleInstruction(const int8_t* name, int16_t offset);
-static int16_t constantInstruction(const int8_t* name, Chunk* chunk, int16_t offset);
-static int16_t longConstantInstruction(const int8_t* name, Chunk* chunk, int16_t offset);
-static int16_t byteInstruction(const int8_t* name, Chunk* chunk, int16_t offset);
-static int16_t jumpInstruction(const int8_t* name, int16_t sign, Chunk* chunk, int16_t offset);
+int32_t disassembleInstruction(Chunk* chunk, int32_t offset);
+static int32_t simpleInstruction(const int8_t* name, int32_t offset);
+static int32_t constantInstruction(const int8_t* name, Chunk* chunk, int32_t offset);
+static int32_t longConstantInstruction(const int8_t* name, Chunk* chunk, int32_t offset);
+static int32_t byteInstruction(const int8_t* name, Chunk* chunk, int32_t offset);
+static int32_t jumpInstruction(const int8_t* name, int32_t sign, Chunk* chunk, int32_t offset);
 void printValue(Value value);
 
 #endif
